requirements.txt
azure-functions==1.23.0
torch==2.4.0
torchvision==0.19.0
pillow==10.4.0
numpy==1.26.4
azure-storage-blob==12.23.0
hydra-core==1.3.2
tqdm==4.66.5